firebase wip: Split frame impls out into files
