+++ backend-firebase/src/firestore/frames/index.js
@@ -1,0 +1,10 @@
+exports.BaseFrame = require('./BaseFrame.js');
+exports.PrimitiveFrame = require('./PrimitiveFrame.js');
+exports.MapFrame = require('./MapFrame.js');
+exports.ListFrame = require('./ListFrame.js');
+exports.BlobFrame = require('./BlobFrame.js');
+exports.PartitionedLogFrame = require('./PartitionedLogFrame.js');
+exports.LogPartitionFrame = require('./LogPartitionFrame.js');
+exports.DocumentFrame = require('./DocumentFrame.js');
+exports.CollectionFrame = require('./CollectionFrame.js');
+exports.RootFrame = require('./RootFrame.js');
